Add EmbeddingGemma-300m embedding support (Gemma3 bidirectional encoder) (#61)
* Support EmbeddingGemma MLX embeddings

* Add EmbeddingGemma-300m embedding support (Gemma3 encoder)

Serves google/embeddinggemma-300m (mlx-community 8-bit) as a bidirectional
encoder + mean-pooling embedding model. Verified numerically against the
mlx-embeddings reference (cosine 0.9996-0.9999 on the mlx-community weights,
batched with padding).

Pipeline (grounded in mlx-lm gemma3_text.py + mlx-embeddings): Gemma3 backbone
(sandwich norms, (1+w) RMSNorm, q/k-norm, dual-RoPE by sliding/global layer,
query_pre_attn_scalar scale, sqrt(hidden) embed scale, GeGLU) with bidirectional
padding-masked attention -> masked mean pool -> Dense(768->3072) -> Dense(3072->768)
(bias-free, identity) -> L2 normalize.

- convert/: map HF `gemma3_text` -> family `embeddinggemma`; dual-RoPE
  (rope_theta / rope_local_base_freq), layer_types, hidden_states_scale=sqrt(h),
  weight_sanitize=HfNormOnly (mlx-community ships raw Gemma gamma norms; lift +1
  at load), and a tensor-map for the two Dense projections.
- core/model.rs: EmbeddingDense0/1 roles; manifest validation skips lm_head and
  shape-checks the Dense head (8-bit packed cols).
- mlx/weights.rs: load the Dense projections; placeholder lm_head (encoder).
- mlx/model: dedicated bidirectional Gemma3 encoder forward + Dense head +
  apply_embedding_dense_head; build_bidirectional_padding_mask materializes the
  mask before its host buffer is freed (fixes a use-after-free -> NaN).
- mlx/runner: route embeddinggemma to the encoder forward, force Mean pooling,
  apply the Dense head; allowlist the family in the MLX support + interleaved-
  attention gates; skip the generation JIT warmup (no decode path).
- bin/embed_gemma_verify.rs: numeric verification harness vs reference vectors.

Usage: `ax-engine-bench generate-manifest <hf_dir>` then
`Session(model_id="embeddinggemma", mlx=True, mlx_model_artifacts_dir=<hf_dir>)
.embed_batch_flat_bytes(batch, pooling="mean", normalize=True)`.


---------

+++ crates/ax-engine-mlx/src/model/families/standard.rs
@@ -848,9 +848,9 @@ pub(crate) fn layer_forward_bidirectional(
         let prompt_start = token_offset.saturating_sub(window);
         if prompt_start > 0 && prompt_start < cached_seq {
             // Slice the prompt KV along the sequence dimension (axis 2).
-            let b = cached_k_full.shape()[0] as i32;
-            let h = cached_k_full.shape()[1] as i32;
-            let d = cached_k_full.shape()[3] as i32;
+            let b = cached_k_full.shape()[0];
+            let h = cached_k_full.shape()[1];
+            let d = cached_k_full.shape()[3];
             let sliced_k = slice(
                 &cached_k_full,
                 &[0, 0, prompt_start as i32, 0],
